FindFood/requirements.txt
Flask
pandas
psycopg2-binary
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-WTF==1.0.1
Werkzeug==2.1.2
WTForms==3.0.1
bcrypt==3.2.2
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.31
psycopg2-binary==2.9.3